True vectorial search & JIT index load
Vectorial search was incorrect, it's now fixed.
(sum on doc words was only partial)

JIT index load: the main part of the index (the reversed index) is loaded when
Load() is called, but all the optional parts are only loaded when they should
be used. It avoids loading parts needing for vectorial search for a boolean
search, and the other way around.